scope test to linux only

+++ test/watcher.js
@@ -472,22 +472,24 @@ describe('watcher', () => {
           assert.deepEqual(res, [{type: 'delete', path: f1}]);
         });
 
-        it('should not crash with rapid folder creates and deletes', async () => {
-          let f1 = getFilename();
-          let f2 = getFilename();
-
-          for (let i = 0; i < 1000; i++) {
-            await fs.mkdir(f1);
-            await fs.rmdir(f1);
-          }
-
-          await nextEvent();
-
-          fs.writeFile(f2, 'hello world');
-
-          let res = await nextEvent();
-          assert.deepEqual(res, [{ type: 'create', path: f2 }]);
-        });
+        if (process.platform === 'linux') {
+          it('linux: should not crash with rapid folder creates and deletes', async () => {
+            let f1 = getFilename();
+            let f2 = getFilename();
+  
+            for (let i = 0; i < 1000; i++) {
+              await fs.mkdir(f1);
+              await fs.rmdir(f1);
+            }
+  
+            await nextEvent();
+  
+            fs.writeFile(f2, 'hello world');
+  
+            let res = await nextEvent();
+            assert.deepEqual(res, [{ type: 'create', path: f2 }]);
+          });
+        }
       });
 
       describe('ignore', () => {
